Refactor both update and create Lambda requests
