+++ src/main/scala/com/gilt/aws/lambda/AwsLambda.scala
@@ -12,21 +12,13 @@ import sbt._
 import scala.util.{Failure, Success, Try}
 
 private[lambda] object AwsLambda {
-  def updateLambda(region: Region, lambdaName: LambdaName, bucketId: S3BucketId, s3Key: S3Key): Try[UpdateFunctionCodeResult] = {
+
+  def updateLambdaWithFunctionCodeRequest(region: Region, lambdaName: LambdaName, updateFunctionCodeRequest: UpdateFunctionCodeRequest): Try[UpdateFunctionCodeResult] = {
     try {
       val client = new AWSLambdaClient(AwsCredentials.provider)
       client.setRegion(RegionUtils.getRegion(region.value))
 
-      val request = {
-        val r = new UpdateFunctionCodeRequest()
-        r.setFunctionName(lambdaName.value)
-        r.setS3Bucket(bucketId.value)
-        r.setS3Key(s3Key.value)
-
-        r
-      }
-
-      val updateResult = client.updateFunctionCode(request)
+      val updateResult = client.updateFunctionCode(updateFunctionCodeRequest)
 
       println(s"Updated lambda ${updateResult.getFunctionArn}")
       Success(updateResult)
@@ -36,56 +28,6 @@ private[lambda] object AwsLambda {
                  _ : AmazonServiceException) =>
         Failure(ex)
     }
-  }
-
-  def updateLambdaWithJar(region: Region, lambdaName: LambdaName, jar: File): Try[UpdateFunctionCodeResult] = {
-    try {
-      val client = new AWSLambdaClient(AwsCredentials.provider)
-      client.setRegion(RegionUtils.getRegion(region.value))
-
-      val request = {
-        val r = new UpdateFunctionCodeRequest()
-        r.setFunctionName(lambdaName.value)
-
-        val buffer = getJarBuffer(jar)
-        r.setZipFile(buffer)
-
-        r
-      }
-
-      val updateResult = client.updateFunctionCode(request)
-
-      println(s"Updated lambda ${updateResult.getFunctionArn}")
-      Success(updateResult)
-    }
-    catch {
-      case ex @ (_ : AmazonClientException |
-                 _ : AmazonServiceException) =>
-        Failure(ex)
-    }
-  }
-
-  def createLambda(region: Region,
-                   jar: File,
-                   functionName: LambdaName,
-                   handlerName: HandlerName,
-                   roleName: RoleARN,
-                   s3BucketId: S3BucketId,
-                   timeout:  Option[Timeout],
-                   memory: Option[Memory],
-                   deployMethod: String
-                  ): Try[CreateFunctionResult] = {
-    val functionCode = new FunctionCode
-
-    if (deployMethod == "S3") {
-        functionCode.setS3Bucket(s3BucketId.value)
-        functionCode.setS3Key(jar.getName)
-    } else {
-        val buffer = getJarBuffer(jar)
-        functionCode.setZipFile(buffer)
-    }
-
-    createLambdaWithFunctionCode(region, jar, functionName, handlerName, roleName, timeout, memory, functionCode)
   }
 
   def createLambdaWithFunctionCode(region: Region,
@@ -124,15 +66,5 @@ private[lambda] object AwsLambda {
                  _ : AmazonServiceException) =>
         Failure(ex)
     }
-  }
-
-  def getJarBuffer(jar: File): ByteBuffer = {
-    val buffer = ByteBuffer.allocate(jar.length().toInt)
-    val aFile = new RandomAccessFile(jar, "r")
-    val inChannel = aFile.getChannel()
-    while (inChannel.read(buffer) > 0) {}
-    inChannel.close()
-    buffer.rewind()
-    buffer
   }
 }
